backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pydantic==1.10.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.5.0
bcrypt==4.0.1

